apply suggestions from code review.
